+++ sim_rl/mujoco_playground/mujoco_playground/_src/manipulation/__init__.py
@@ -53,6 +53,7 @@ _envs = {
     "AeroCubeRotateZAxis": aero_hand_rotate_z.CubeRotateZAxis,
     "AeroCubeGraspV2Force": aero_hand_grasp_cube_v2_force.CubeGraspV2Force,
     "AeroCubeGraspV2ForceCoacd": aero_hand_grasp_cube_v2_force.CubeGraspV2ForceCoacd,
+    "AeroCubeGraspV2ForceCoacdQbr": aero_hand_grasp_cube_v2_force.CubeGraspV2ForceCoacdQbr,
     "AeroBottleGraspV2Force": aero_hand_grasp_bottle_v2_force.BottleGraspV2Force,
     "AeroCanGraspV2Force": aero_hand_grasp_can_v2_force.CanGraspV2Force,
 }
@@ -75,6 +76,7 @@ _cfgs = {
     "AeroCubeRotateZAxis": aero_hand_rotate_z.default_config,
     "AeroCubeGraspV2Force": aero_hand_grasp_cube_v2_force.default_config,
     "AeroCubeGraspV2ForceCoacd": aero_hand_grasp_cube_v2_force.default_config,
+    "AeroCubeGraspV2ForceCoacdQbr": aero_hand_grasp_cube_v2_force.default_config_qbr,
     "AeroBottleGraspV2Force": aero_hand_grasp_bottle_v2_force.default_config,
     "AeroCanGraspV2Force": aero_hand_grasp_can_v2_force.default_config,
 }
@@ -90,6 +92,7 @@ _randomizer = {
     "AeroCubeRotateZAxis": aero_hand_rotate_z.domain_randomize,
     "AeroCubeGraspV2Force": aero_hand_grasp_cube_v2_force.domain_randomize,
     "AeroCubeGraspV2ForceCoacd": aero_hand_grasp_cube_v2_force.domain_randomize,
+    "AeroCubeGraspV2ForceCoacdQbr": aero_hand_grasp_cube_v2_force.domain_randomize,
     "AeroBottleGraspV2Force": aero_hand_grasp_bottle_v2_force.domain_randomize,
     "AeroCanGraspV2Force": aero_hand_grasp_can_v2_force.domain_randomize,
 }