+++ sim_rl/mujoco_playground/mujoco_playground/_src/manipulation/aero_hand/aero_hand_constants.py
@@ -110,6 +110,7 @@ SENSOR_HW6_FORCE_NAMES = [
 
 GRASP_V2_XML = ROOT_PATH / "xmls" / "scene_mjx_grasp_v2.xml"
 GRASP_V2_COACD_XML = ROOT_PATH / "xmls" / "scene_mjx_grasp_v2_coacd.xml"
+GRASP_V2_COACD_QBR_XML = ROOT_PATH / "xmls" / "scene_mjx_grasp_v2_coacd_qbr.xml"
 GRASP_V2_BOTTLE_XML = ROOT_PATH / "xmls" / "scene_mjx_grasp_bottle_550ml.xml"
 GRASP_V2_CAN_XML = ROOT_PATH / "xmls" / "scene_mjx_grasp_can_330ml.xml"